cmake cleanup

+++ src/lib/controllib/BlockRandGauss.hpp
@@ -85,10 +85,10 @@ public:
 				S = V1 * V1 + V2 * V2;
 			} while (S >= 1 || fabsf(S) < 1e-8f);
 
-			X = V1 * float(sqrtf(-2 * float(logf(S)) / S));
+			X = V1 * sqrtf(-2 * logf(S) / S);
 
 		} else {
-			X = V2 * float(sqrtf(-2 * float(logf(S)) / S));
+			X = V2 * sqrtf(-2 * logf(S) / S);
 		}
 
 		phase = 1 - phase;
